illumination: Add common interface for directional-like illuminants
This commit adds a common interface for the DirectionalIllumination and
AstroObjectIllumination class. This makes sense, since it is strange to
say that "an AstroObjectIllumination is a DirectionalIllumination".

In addition, we fix the azimuth field init sequence to allow all values
and warn if the user inputs something out of the expected range.
