+++ docs/src/release_notes/v0.27.x.md
@@ -33,6 +33,8 @@
   ({ghpr}`397`).
 * 📖 The Data documentation content is extracted from the user guide and moved
   to its own section ({ghpr}`405`).
+* Illumination azimuth values out of the [0°, 360°[ range are now allowed and
+  issue a warning instead of raising an exception ({ghpr}`409`).
 
 ### Fixed
