+++ src/eradiate/scenes/illumination/__init__.pyi
@@ -1,5 +1,6 @@
 from ._astro_object import AstroObjectIllumination as AstroObjectIllumination
 from ._constant import ConstantIllumination as ConstantIllumination
+from ._core import AbstractDirectionalIllumination as AbstractDirectionalIllumination
 from ._core import Illumination as Illumination
 from ._core import illumination_factory as illumination_factory
 from ._directional import DirectionalIllumination as DirectionalIllumination
